+++ src/main/kotlin/spock/adb/command/ClearAppDataAndRestartCommand.kt
@@ -2,7 +2,6 @@ package spock.adb.command
 
 import com.android.ddmlib.IDevice
 import com.intellij.openapi.project.Project
-
 import spock.adb.clearAppData
 import spock.adb.getDefaultActivityForApplication
 import spock.adb.isAppInstall