+++ src/main/kotlin/spock/adb/command/WindowAnimatorScaleCommand.kt
@@ -2,8 +2,9 @@ package spock.adb.command
 
 import com.android.ddmlib.IDevice
 import com.intellij.openapi.project.Project
-import java.util.concurrent.TimeUnit
 import spock.adb.ShellOutputReceiver
+import spock.adb.ShellQuote
+import java.util.concurrent.TimeUnit
 
 class WindowAnimatorScaleCommand : Command<String, String> {
 
@@ -14,7 +15,7 @@ class WindowAnimatorScaleCommand : Command<String, String> {
     override fun execute(p: String, project: Project, device: IDevice): String {
         val shellOutputReceiver = ShellOutputReceiver()
         device.executeShellCommand(
-            "settings put global window_animation_scale $p",
+            "settings put global window_animation_scale ${ShellQuote.quote(p)}",
             shellOutputReceiver,
             15L,
             TimeUnit.SECONDS